workflows/requirements.txt
PyGithub
PyYAML
glob2
